Windows compatibility and instructions for pybonjour on Python 3.7

+++ pavement.py
@@ -7,6 +7,7 @@ from paver.easy import *
 from paver.setuputils import setup
 import urllib.request
 import os
+import platform
 
 setup(
     name="TestProject",
@@ -19,20 +20,29 @@ setup(
 
 @task
 def libunrar():
-    UNRAR_SRC = "http://www.rarlab.com/rar/unrarsrc-5.3.2.tar.gz"
+    if platform.system() == "Windows":
+        UNRAR_SRC = "http://www.rarlab.com/rar/UnRARDLL.exe"
+    else:
+        UNRAR_SRC = "http://www.rarlab.com/rar/unrarsrc-5.3.2.tar.gz"
+
     dest = "libunrar/" + os.path.basename(UNRAR_SRC)
+
     if not os.path.exists(dest):
         print("Fetching " + UNRAR_SRC + "...")
         urllib.request.urlretrieve(UNRAR_SRC, dest)
-    if not os.path.exists("libunrar/unrar"):
-        print("Unpacking " + dest + "...")
-        sh("tar xfz %s -C libunrar" % (dest))
-    print("Compiling unrar...")
-    sh("cd libunrar/unrar && make lib")
-    sh("cp libunrar/unrar/libunrar.so libunrar/libunrar.so")
-    print("Cleaning up ...")
-    sh("rm -rf libunrar/unrar*")
-    sh("rm -f libunrar/unrar/libunrar.so")
+
+    if platform.system() == "Windows":
+        print("Please extract UnRAR.dll from UnRARDLL.exe into libunrar/libunrar.so")
+    else:
+        if not os.path.exists("libunrar/unrar"):
+            print("Unpacking " + dest + "...")
+            sh("tar xfz %s -C libunrar" % (dest))
+        print("Compiling unrar...")
+        sh("cd libunrar/unrar && make lib")
+        sh("cp libunrar/unrar/libunrar.so libunrar/libunrar.so")
+        print("Cleaning up ...")
+        sh("rm -rf libunrar/unrar*")
+        sh("rm -f libunrar/unrar/libunrar.so")
 
 
 @task
